Handle SageMaker API limitation of 2500 chars in payload
